==> 0.Using_EMI_formula.ipynb <==
### In this notebook I have tried to calculate the ROI(Rate of intrest) using the EMI formula 
### Estimations and guess made for missing values are following:
#### 1. *ACTUAL_PAYMT_AMT=EMI_AMOUNT*
#### 2. *Assigned random values to repayment tenure between a range using domain specicifc knowledge*

import pandas as pd


df=pd.read_csv('train_main_loan.csv')

df = df.dropna(subset=['ACTUAL_PAYMT_AMT'])


df['EMI_AMOUNT'].fillna(df['ACTUAL_PAYMT_AMT'], inplace=True)


df['PRINCIPAL'] = df['CURRENT_BALANCE'] + df['ACTUAL_PAYMT_AMT']


import numpy as np

# Assuming your DataFrame is named 'df'

# Define functions to generate random values within the specified ranges
def generate_random_personal_loan():
    return np.random.randint(45,46)

def generate_random_business_loan():
    return np.random.randint(48,50)

def generate_random_housing_loan():
    return np.random.randint(175,176)

def generate_random_property_loan():
    return np.random.randint(125, 126)

# Create a mapping dictionary for each 'ACCOUNT_TYPE' to the respective generator function
repayment_tenure_mapping = {
    'Personal Loan': generate_random_personal_loan,
    'Business Loan': generate_random_business_loan,
    'Housing Loan': generate_random_housing_loan,
    'Property Loan': generate_random_property_loan
}

# Fill missing values in 'REPAYMENT_TENURE' based on 'ACCOUNT_TYPE'
for account_type, generator_function in repayment_tenure_mapping.items():
    mask = (df['ACCOUNT_TYPE'] == account_type) & (df['REPAYMENT_TENURE'].isna())
    df.loc[mask, 'REPAYMENT_TENURE'] = df.loc[mask].apply(lambda x: generator_function(), axis=1)

# Now, 'df' contains filled values in 'REPAYMENT_TENURE' based on 'ACCOUNT_TYPE'


new_df = df[['ID','PRINCIPAL', 'EMI_AMOUNT', 'REPAYMENT_TENURE', 'ACTUAL_ROI']]
new_df.to_csv('domain_knowledge.csv')

df=pd.read_csv('domain_knowledge.csv')

df

import pandas as pd
import numpy as np
from sklearn.metrics import r2_score

# Assuming your DataFrame is named 'df'
df = df[~df['CALCULATED_ROI'].astype(str).str.contains('#NUM!')]

# Convert 'ACTUAL_ROI' and 'CALCULATED_ROI' columns to numeric, handling errors with 'coerce'
df['CALCULATED_ROI'] = pd.to_numeric(df['CALCULATED_ROI'], errors='coerce')

# Drop rows with NaN values in the 'CALCULATED_ROI' column
df.dropna(subset=['CALCULATED_ROI'], inplace=True)

# Filter rows based on conditions
df = df[(df['ACTUAL_ROI'] >= 0) & (df['CALCULATED_ROI'] >= 0) & (df['CALCULATED_ROI'] <= 0.50)]

# Drop rows with the string '#NUM!' in 'ACTUAL_ROI' and 'CALCULATED_ROI'

# Calculate R2 between 'ACTUAL_ROI' and 'CALCULATED_ROI'
actual_roi = df['ACTUAL_ROI']/100
calculated_roi = df['CALCULATED_ROI']
r2_value = r2_score(actual_roi, calculated_roi)

# Now, 'df' contains the filtered rows, and you have the R2 value


df

print(r2_value)
