==> 5.Feature_selection.ipynb <==
### In this notebook PCA (Principal Component Analysis) and the checking the relationship betweeen traget varibles and PCA features using MI (Mutal Information)

import pandas as pd
df_train=pd.read_csv("feature_generated.csv")

import pandas as pd
from sklearn.decomposition import PCA
from sklearn.preprocessing import StandardScaler
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline

# Assuming df_train is your DataFrame
df = df_train.copy()
df = df.drop(columns=['ACTUAL_ROI', 'ID', 'PAYMENT_HISTORY_1'])

# Separate numerical and categorical features
numerical_features = df.select_dtypes(include=['float64', 'int64']).columns

# Create transformer for numerical features
numerical_transformer = StandardScaler()

# Combine transformer
preprocessor = ColumnTransformer(
    transformers=[
        ('num', numerical_transformer, numerical_features)
    ])

pca = PCA(n_components=5)  # Set n_components to 10 for top 10 features
pipeline = Pipeline(steps=[('preprocessor', preprocessor), ('pca', pca)])
X_pca = pipeline.fit_transform(df)
feature_names = list(preprocessor.named_transformers_['num'].get_feature_names_out(numerical_features))
top_10_feature_names = [f'{feature_names[i]}' for i in range(5)]
df_pca = pd.DataFrame(data=X_pca, columns=top_10_feature_names)
print(top_10_feature_names)

feature_names

numerical_df = df_train.select_dtypes(exclude=['O'])
X = numerical_df.drop(['ACTUAL_ROI'], axis=1)
y = numerical_df['ACTUAL_ROI']

import pandas as pd
from sklearn.feature_selection import mutual_info_regression

mi_scores = mutual_info_regression(X, y)

mi_df = pd.DataFrame({'Feature': X.columns, 'MI Score': mi_scores})

mi_df_sorted = mi_df.sort_values(by='MI Score', ascending=False)

# Display the sorted DataFrame
print(mi_df_sorted)
